+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### Added
 - ReportingApi with attachment support
+- `extendTestDescriptionWithLastError` option to the RP config to be able to toggle the last error log attaching to the test description.
 
 ## [5.0.0] - 2024-02-15
 ### Added